+++ src/components/select/filterable-select/filterable-select.component.js
@@ -71,15 +71,20 @@ const FilterableSelect = React.forwardRef(
 
     const updateValues = useCallback(
       (newFilterText, isDeleteEvent) => {
-        setFilterText(newFilterText);
-
         setSelectedValue((previousValue) => {
           const match = findElementWithMatchingText(newFilterText, children);
+          const isFilterCleared = isDeleteEvent && newFilterText === "";
 
-          if (!match || isDeleteEvent) {
+          if (!match || isFilterCleared) {
             setTextValue(newFilterText);
 
-            return previousValue;
+            return "";
+          }
+
+          if (isDeleteEvent) {
+            setTextValue(newFilterText);
+
+            return match.props.value;
           }
 
           if (onChange) {
@@ -109,16 +114,23 @@ const FilterableSelect = React.forwardRef(
     );
 
     const setMatchingText = useCallback(
-      (newValue) => {
+      (newValue, isClosing) => {
         const matchingOption = React.Children.toArray(children).find((child) =>
           isExpectedOption(child, newValue)
         );
 
-        if (matchingOption && matchingOption.props.text !== undefined) {
+        if (!matchingOption || matchingOption.props.text === undefined) {
+          setTextValue("");
+        } else if (
+          isClosing ||
+          matchingOption.props.text
+            .toLowerCase()
+            .startsWith(filterText.toLowerCase())
+        ) {
           setTextValue(matchingOption.props.text);
         }
       },
-      [children]
+      [children, filterText]
     );
 
     const handleTextboxChange = useCallback(
@@ -130,6 +142,7 @@ const FilterableSelect = React.forwardRef(
           event.nativeEvent.inputType === "delete";
 
         updateValues(newValue, isDeleteEvent);
+        setFilterText(newValue);
         setOpen(true);
       },
       [updateValues]
@@ -183,7 +196,7 @@ const FilterableSelect = React.forwardRef(
         isMouseDownReported.current = false;
 
         if (notInContainer && notInList) {
-          setMatchingText(selectedValue);
+          setMatchingText(selectedValue, true);
           setOpen(false);
         }
       },
@@ -208,16 +221,17 @@ const FilterableSelect = React.forwardRef(
       );
 
       setSelectedValue((prevValue) => {
-        if (isControlled.current && prevValue && !newValue) {
-          setFilterText("");
-          setTextValue("");
+        if (isControlled.current && prevValue !== newValue) {
+          setMatchingText(newValue);
         }
 
         return newValue;
       });
 
       setHighlightedValue(newValue);
-    }, [value, defaultValue, onChange]);
+      // prevent value update on filter change
+      // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [value, defaultValue, onChange, children]);
 
     useEffect(() => {
       if (!isOpen) {
@@ -240,8 +254,9 @@ const FilterableSelect = React.forwardRef(
 
     useEffect(() => {
       setMatchingText(value || defaultValue);
+      // update text value only when children are changing
       // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
+    }, [value, children]);
 
     useEffect(() => {
       const clickEvent = "click";
@@ -302,7 +317,7 @@ const FilterableSelect = React.forwardRef(
 
     const onSelectListClose = useCallback(() => {
       setOpen(false);
-      setMatchingText(selectedValue);
+      setMatchingText(selectedValue, true);
     }, [selectedValue, setMatchingText]);
 
     function handleTextboxClick(event) {